chore: add unit tests for `getCommonBase()`
