+++ src/transform.ts
@@ -2,10 +2,9 @@ import { basename, posix } from 'path'
 import MagicString from 'magic-string'
 import fg from 'fast-glob'
 import { stringifyQuery } from 'ufo'
-import { scan } from 'micromatch'
 import type { PluginOptions } from '../types'
 import { parseImportGlob } from './parse'
-import { assert, isCSSRequest } from './utils'
+import { assert, getCommonBase, isCSSRequest } from './utils'
 
 const importPrefix = '__vite_glob_next_'
 
@@ -41,7 +40,7 @@ export async function transform(
 
   const staticImports = (await Promise.all(
     matches.map(async({ globsResolved, isRelative, options, index, start, end }) => {
-      const cwd = getCommonAncestor(globsResolved) ?? root
+      const cwd = getCommonBase(globsResolved) ?? root
       const files = (await fg(globsResolved, {
         cwd,
         absolute: true,
@@ -136,21 +135,4 @@ export async function transform(
     s,
     matches,
   }
-}
-
-function getCommonAncestor(globsResolved: string[]): null | string {
-  const bases = globsResolved.filter(g => !g.startsWith('!')).map(glob => scan(glob).base)
-  if (!bases.length)
-    return null
-
-  let commonAncestor = '/'
-  const dirS = bases[0].split('/')
-  for (let i = 0; i < dirS.length; i++) {
-    const candidate = dirS.slice(0, i + 1).join('/')
-    if (bases.every(base => base.startsWith(candidate)))
-      commonAncestor = candidate
-    else
-      break
-  }
-  return commonAncestor
 }
