+++ src/utils.ts
@@ -1,3 +1,5 @@
+import { scan } from 'micromatch'
+
 const cssLangs = '\\.(css|less|sass|scss|styl|stylus|pcss|postcss)($|\\?)'
 const cssLangRE = new RegExp(cssLangs)
 
@@ -14,4 +16,21 @@ export function assert(condition: unknown): asserts condition {
       'Reach out at https://github.com/antfu/vite-plugin-glob/issues/new and include this error stack (the error stack is usually enough to fix the problem).',
     ].join(' '),
   )
+}
+
+export function getCommonBase(globsResolved: string[]): null | string {
+  const bases = globsResolved.filter(g => !g.startsWith('!')).map(glob => scan(glob).base)
+  if (!bases.length)
+    return null
+
+  let commonAncestor = '/'
+  const dirS = bases[0].split('/')
+  for (let i = 0; i < dirS.length; i++) {
+    const candidate = dirS.slice(0, i + 1).join('/')
+    if (bases.every(base => base.startsWith(candidate)))
+      commonAncestor = candidate
+    else
+      break
+  }
+  return commonAncestor
 }
